Move test code to test_ping directory
We will add more tests.
