+++ remote-trait-object-tests/src/test_ping/mod_main/context.rs
