+++ remote-trait-object-tests/src/test_ping/mod_ping/traits.rs
